# packages/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Loombre API
  version: 1.0.0
  summary: >-
    Contract-first API for Loombre, a ground-up self-hosted media streaming
    platform (movies, TV, music).
  description: >-
    This document is the source of truth (CLAUDE.md architecture invariant
    #1): apps/server controllers conform to it and packages/sdk is generated
    from it — never hand-authored.


    Conventions (apply globally; see components/ for shared shapes): all
    timestamps are BIGINT epoch milliseconds as JSON integers with field
    names suffixed `Ms`; all ids are UUIDv7 strings; all list endpoints use
    cursor pagination (`{ items, nextCursor }`, with `?cursor=&limit=` query
    params); all 4xx/5xx responses are RFC 9457 `application/problem+json`;
    enums are closed OpenAPI enums; core entity schemas set
    `additionalProperties: false`.


    Evolution policy: additive-only within a major version (docs/PLAN.md
    §4.1) — CI runs oasdiff against `main` and fails the build on any
    breaking change classification. Deprecated operations carry
    `deprecated: true` plus a `Sunset` response header for two minor releases
    minimum before a major bump.
  license:
    name: AGPL-3.0-relicense-pending
    identifier: LicenseRef-Loombre-Private
...
servers:
  - url: /v1
    description: Loombre API, version 1 (path-prefixed; host is deployment-specific)
security:
  - bearerAuth: []
...
paths:
...
  /admin/settings/{key}:
    put:
      operationId: updateAdminSetting
      tags: [admin]
      summary: Update one UI-editable setting's value (admin)
      description: >-
        STATE.md Addendum A decision A10: re-verifies the caller's admin
        status with a FRESH database read at mutation time (never trusted
        from a possibly-stale access-token claim) before anything else,
        including before revealing whether `key` exists at all. Ordered
        checks: 403 (not currently an admin, live re-verify) -> 404 (unknown
        key, or a scope:'env-only' key — those are never writable through
        this surface, the lockout boundary docs/PLAN.md describes) -> 409
        (an active env pin governs this key right now; the submitted value
        is discarded unconditionally, valid or not) -> 422 (schema
        validation, including the restricted.majorityAgeYears >=18 floor,
        which cannot be configured below it through any path).
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z][a-zA-Z0-9]*(\.[a-zA-Z][a-zA-Z0-9]*)+$'
          description: >-
            A scope:'ui' registry key exactly as listed by GET
            /admin/settings/schema (e.g.
            "transcode.maxSimultaneousTranscodes"). Dotted
            category.entryName form, matching every entry in
            packages/shared/src/settings-registry.ts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSettingRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UpdateSettingResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            The key is currently pinned by its environment variable (A8) —
            env always wins; the submitted value cannot take effect while
            the pin is active, regardless of whether it would otherwise be
            valid on its own. The DB value underneath, if any, is preserved
            but left inert.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }

  # ──────── Admin: provider API keys (Addendum A, decisions A6/A9) ──────────
  # Deliberately NOT nested under /admin/settings/{key}: a
  # /admin/settings/provider-keys/{provider} path would sit one segment away
  # from being swallowed by the {key} param above — the same
  # /users/{id}-shadowed-by-/users/me class of trap (Phase 1). Keys are
  # secrets, never configuration (provider-keys.service.ts's header) — GET
  # never exists for a key VALUE anywhere in this contract; status rides on
  # GET /admin/settings's `providerKeys` only.
  /admin/provider-keys/{provider}:
    parameters:
      - name: provider
        in: path
        required: true
        schema: { $ref: '#/components/schemas/ProviderName' }
    put:
      operationId: setAdminProviderKey
      tags: [admin]
      summary: Set (or replace) a provider API key (admin, write-only)
      description: >-
        A9: write-only — the submitted key is never echoed back, logged, or
        otherwise readable again through this API (status reads — GET
        /admin/settings — report set/source/lastSetMs only, never the
        value). Stored via the platform secret keyring (packages/secrets),
        never in server_settings (no encryption-at-rest story there). A10
        live-admin re-verify gates this exactly like updateAdminSetting.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetProviderKeyRequest' }
      responses:
        '204': { description: Key stored }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        default: { $ref: '#/components/responses/Problem' }
    delete:
      operationId: clearAdminProviderKey
      tags: [admin]
      summary: Clear a provider API key (admin)
      description: >-
        Removes the stored keyring entry. A real env var for this provider
        (LOOMBRE_TMDB_API_KEY / LOOMBRE_TVDB_API_KEY), if set, still wins on
        the next status read regardless (A8 precedence) — this only clears
        the keyring-stored fallback.
      responses:
        '204': { description: Key cleared }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }

  # ──────── Admin: plugins (LPP v1, Lane W5 — packages/plugin-protocol/spec/
  # lpp-v1.md is the frozen wire contract this surface administers;
  # apps/server/src/plugins/* holds the registration/lifecycle/health
  # services this is a thin admin API over). Every operation re-verifies
  # admin LIVE (A10 pattern) — plugins are a security surface (out-of-process
  # HTTP services holding a delivery-signing secret and, per capability,
  # visibility into library metadata or the activity feed). Response schemas
  # never carry a secret value: hmacSecret is returned exactly once, by
  # registerAdminPlugin and rotateAdminPluginHmac only; `config` on AdminPlugin
  # holds non-secret configSchema values only (LD1) — a plugin's own
  # manifest.configSchema tells the UI which field names are secret: true and
  # therefore keyring-only. ──────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access JWT; refreshed via per-device rotating refresh tokens.
...
  responses:
    Problem:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found (or not visible under the caller's ViewerContext)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableEntity:
      description: Request failed schema/business-rule validation
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Extension members are additive.
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer, minimum: 100, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code:
          type: string
          description: Stable machine-readable error code, additional to `type`.
      required: [title, status]
      additionalProperties: true
...
    SettingsValueSource:
      type: string
      enum: [environment, database, default]
      description: Which input actually determined a setting's current effective value (A8 precedence order).
...
    ProviderName:
      type: string
      enum: [tmdb, tvdb]
      description: Closed set of metadata providers with an admin-manageable API key (A9).
...
    UpdateSettingRequest:
      type: object
      additionalProperties: false
      required: [value]
      properties:
        value: {}
    UpdateSettingResponse:
      type: object
      additionalProperties: false
      required: [key, value, source, requiresRestart, restartPending]
      properties:
        key: { type: string }
        value: {}
        source: { $ref: '#/components/schemas/SettingsValueSource' }
        requiresRestart: { type: boolean }
        restartPending:
          type: boolean
          description: Whether THIS key now appears in restartPendingKeys after this write — lets the UI show "restart required" inline without a second GET.
...
    SetProviderKeyRequest:
      type: object
      additionalProperties: false
      required: [key]
      properties:
        key:
          type: string
          minLength: 1
          description: The raw provider API key. Write-only — never returned by any endpoint.

    # ═══════════════════ Admin: plugins (LPP v1, Lane W5) ═══════════════════
...